+++ src/main/java/lotto/domain/Lotto.java
@@ -4,7 +4,7 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import lotto.view.constant.ExceptionMessage;
+import lotto.exception_message.ExceptionMessage;
 
 public class Lotto {
