+++ src/main/java/lotto/domain/WinningLotto.java
@@ -1,7 +1,7 @@
 package lotto.domain;
 
 import java.util.List;
-import lotto.view.constant.ExceptionMessage;
+import lotto.exception_message.ExceptionMessage;
 
 public class WinningLotto {
