+++ src/main/java/lotto/exception_message/ExceptionMessage.java
@@ -1,4 +1,4 @@
-package lotto.view.constant;
+package lotto.exception_message;
 
 public enum ExceptionMessage {
     DUPLICATED_NUMBERS("중복되는 번호가 있습니다."),
